Add files via upload
Install and setup

This Notebook allows to read the POLA rates as a time series in ASCII format, plot and compute the Lomb-Scargle periodogram etc. etc.

It was created on Macosx with python3 and X-code. Further packages to be installed:

pip3 install matplotlib # (already including numpy)
pip3 install scipy pandas gatspy
pip3 install notebook
On macosx, run the notebook from a terminal with one of the following:

jupyter notebook
python3 -m notebook
The notebook will open in the browser.